+++ impl/src/utils.rs
@@ -16,6 +16,8 @@ use syn::{
     feature = "add",
     feature = "add_assign",
     feature = "as_ref",
+    feature = "clone",
+    feature = "copy",
     feature = "debug",
     feature = "display",
     feature = "eq",
@@ -46,6 +48,8 @@ pub(crate) use self::generics_search::GenericsSearch;
     feature = "add",
     feature = "add_assign",
     feature = "as_ref",
+    feature = "clone",
+    feature = "copy",
     feature = "debug",
     feature = "display",
     feature = "eq",
@@ -1331,6 +1335,8 @@ pub fn is_type_parameter_used_in_type(
     feature = "as_ref",
     feature = "debug",
     feature = "display",
+    feature = "clone",
+    feature = "copy",
     feature = "eq",
     feature = "from",
     feature = "from_str",
@@ -1408,6 +1414,8 @@ mod either {
     feature = "add",
     feature = "add_assign",
     feature = "as_ref",
+    feature = "clone",
+    feature = "copy",
     feature = "debug",
     feature = "display",
     feature = "eq",
@@ -1513,6 +1521,8 @@ mod spanning {
     feature = "as_ref",
     feature = "debug",
     feature = "display",
+    feature = "clone",
+    feature = "copy",
     feature = "eq",
     feature = "from",
     feature = "from_str",
@@ -1532,6 +1542,14 @@ pub(crate) mod attr {
     };
 
     use super::{Either, Spanning};
+
+    #[cfg(any(
+        feature = "debug",
+        feature = "display",
+        feature = "clone",
+        feature = "copy"
+    ))]
+    pub(crate) use self::bounds::Bounds;
 
     #[cfg(any(
         feature = "as_ref",
@@ -1700,6 +1718,59 @@ pub(crate) mod attr {
                 ))
             })
         }
+    }
+
+    #[cfg(any(
+        feature = "debug",
+        feature = "display",
+        feature = "clone",
+        feature = "copy"
+    ))]
+    mod bounds {
+        use crate::utils::attr::ParseMultiple;
+        use syn::{
+            parse::{Parse, ParseStream},
+            punctuated::Punctuated,
+            spanned::Spanned as _,
+            token,
+        };
+
+        /// Representation of a `bound` macro attribute, expressing additional trait bounds.
+        ///
+        /// ```rust,ignore
+        /// #[<attribute>(bound(<where-predicates>))]
+        /// #[<attribute>(bounds(<where-predicates>))]
+        /// #[<attribute>(where(<where-predicates>))]
+        /// ```
+        #[derive(Debug, Default)]
+        pub struct Bounds(pub Punctuated<syn::WherePredicate, token::Comma>);
+
+        impl Parse for Bounds {
+            fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
+                let _ = input.parse::<syn::Path>().and_then(|p| {
+                    if ["bound", "bounds", "where"]
+                        .into_iter()
+                        .any(|i| p.is_ident(i))
+                    {
+                        Ok(p)
+                    } else {
+                        Err(syn::Error::new(
+                            p.span(),
+                            "unknown attribute argument, expected `bound(...)`",
+                        ))
+                    }
+                })?;
+
+                let content;
+                syn::parenthesized!(content in input);
+
+                content
+                    .parse_terminated(syn::WherePredicate::parse, token::Comma)
+                    .map(Self)
+            }
+        }
+
+        impl ParseMultiple for Bounds {}
     }
 
     #[cfg(any(