+++ tests/compile_fail/clone/union.rs
@@ -1,0 +1,6 @@
+#[derive(derive_more::Clone)]
+union MyUnion {
+    i: u32,
+}
+
+fn main() {}